split bee and digger in 2 files
